+++ packages/api/src/api/users/me/index.ts
@@ -1,0 +1,6 @@
+import { users } from '..';
+import { mountUserMiddleware } from '../../../middleware/mountUserMiddleware';
+
+users.get('/api/users/me', mountUserMiddleware, (req, res) => {
+  res.json(req.user);
+});
